Move column decoders into module
